% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{dict}
\alias{dict}
\title{Diccionario para los datos de la ENFT
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\format{
\link{list}
}
\source{
\url{https://github.com/endomer/enftr/tree/main/Files/Cuestionarios/}
}
\usage{
dict
}
\description{
Contiene las etiquetas tanto para las variables como para los valores de estas.
}
\keyword{datasets}
