% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{enft}
\alias{enft}
\title{Conjunto de datos que simula la encuesta de fuerza de trabajo
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\format{
\link{data.frame}
}
\usage{
enft
}
\description{
Conjunto de datos que simula la encuesta de fuerza de trabajo
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{datasets}
