% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educacion.R
\name{ft_alfabeta}
\alias{ft_alfabeta}
\title{Alfabetismo de la persona (Sabe leer y escribir)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_alfabeta(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} una conexión a base de datos o data.frame con los datos}

\item{min_edad}{\link{integer} la edad mínima para el cálculo de la variable}
}
\value{
\link{data.frame}: Los datos suministrados en el argumento \code{tbl} con
la variable \code{alfabeta} adicionada.
}
\description{
Entre 2000 y 2005, se incluyó cero (0) en la variable para identificar aquellos
casos en que la pregunta no aplicaba. Esta variable omite esos valores.
}
\examples{
\dontrun{
 enft <- ft_alfabeta(enft)
}
}
