% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_compute_ano.R
\name{ft_ano}
\alias{ft_ano}
\alias{ft_compute_ano}
\title{Año}
\usage{
ft_ano(tbl)

ft_compute_ano(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable \code{ano}
adicionada.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Crea la variable año.
}
\examples{
  (enft <- data.frame(EFT_PERIODO = "1/2016"))
  ft_ano(enft)
}
