% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_cantidad_personas_trabajan}
\alias{ft_cantidad_personas_trabajan}
\title{Cantidad de personas trabajan en la empresa
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_cantidad_personas_trabajan(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{cantidad_personas_trabajan} agregada.
}
\description{
Esta función homologa los rangos de cantidad de personas que laboran en la
empresa. Para el período 2000-2003 se incluían solo 4 categorías, pero de
2004 en adelante se incluyen 7 categorías. Vea los cuestionarios
correspondientes a esos periodos para más información.
}
\examples{
\dontrun{
enft <- ft_cantidad_personas_trabajan(enft)
}
}
