% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_categoria_ocupacion_principal}
\alias{ft_categoria_ocupacion_principal}
\title{Categoría de la ocupación principal
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_categoria_ocupacion_principal(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{categoria_ocupacion_principal} agregada.
}
\description{
Las encuestas levantadas entre 2000 y 2004 contemplaban 10 categorías de
ocupación, pero en 2005 se redujeron a 8. Vea los cuestionarios
correspondientes a esos periodos para más información.
}
\details{
Esta función homologa las categorías de ocupación a las categorías de
menor frecuencia en la encuesta. Dígase las utilizadas a partir de 2005.
}
\examples{
\dontrun{
enft <- ft_categoria_ocupacion_principal(enft)
}
}
