% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_compute_ano.R
\name{ft_compute_ano}
\alias{ft_compute_ano}
\title{Año}
\usage{
ft_compute_ano(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable \code{ano}
adicionada.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
}
\details{
Crea la variable año.
}
\examples{
  (enft <- data.frame(EFT_PERIODO = "1/2016"))
  ft_compute_ano(enft)
}
