% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_db_connect.R
\name{ft_db_connect}
\alias{ft_db_connect}
\alias{ft_dbConnect}
\title{Conexión a base de datos
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Vea \code{Dmisc::\link[Dmisc:db_connect]{db_connect}}}
\usage{
ft_db_connect(db_name = "enft", ...)

ft_dbConnect(db_name = "enft", ...)
}
\arguments{
\item{db_name}{Nombre de la base de datos}

\item{...}{Others arguments to pass to \code{Dmisc::\link[Dmisc:db_connect]{db_connect}}}
}
\value{
Conexión a base de datos.
}
\description{
Conexión a base de datos
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Vea \code{Dmisc::\link[Dmisc:db_connect]{db_connect}}
}
\examples{
\dontrun{
  conn <- ft_db_connect()
}
}
