% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_dbConnect.R
\name{ft_dbConnect}
\alias{ft_dbConnect}
\title{Conexión a base de datos}
\usage{
ft_dbConnect()
}
\value{
Conexión a base de datos.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Vea \code{Dmisc::\link[Dmisc:dbConnect]{dbConnect}}
}
\examples{
\dontrun{
  conn <- ft_dbConnect()
}
}
