% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_desempleo_cesante_abierto}
\alias{ft_desempleo_cesante_abierto}
\title{Población censante en condición de desempleo abierto}
\usage{
ft_desempleo_cesante_abierto(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{desempleo_cesante_abierto} agregada.
}
\description{
Población censante en condición de desempleo abierto
}
\examples{
\dontrun{
enft <- ft_desempleo_cesante_abierto(enft)
}
}
