% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_desempleo_cesante_ampliado}
\alias{ft_desempleo_cesante_ampliado}
\title{Población censante en condición de desempleo ampliado}
\usage{
ft_desempleo_cesante_ampliado(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{desempleo_cesante_ampliado} agregada.
}
\description{
Población censante en condición de desempleo ampliado
}
\examples{
\dontrun{
enft <- ft_desempleo_cesante_ampliado(enft)
}
}
