% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_dias_semana_ocupacion_principal}
\alias{ft_dias_semana_ocupacion_principal}
\title{Número de días trabajados a la semana por ocupación principal
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_dias_semana_ocupacion_principal(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{dias_semana_ocupacion_principal} agregada.
}
\description{
Número de días trabajados a la semana por ocupación principal
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
enft <- dias_semana_ocupacion_principal(enft)
}
}
