% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demograficos.R
\name{ft_dominios_inferencia1}
\alias{ft_dominios_inferencia1}
\title{Dominios de inferencia ENFT (20001 - 20031)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_dominios_inferencia1(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} con datos de la base de datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{dominios_inferencia1} adicionada.
}
\description{
Dominios de inferencia ENFT (20001 - 20031)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
enft <- ft_dominios_inferencia1(enft)
}
}
