% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_grupo_rama}
\alias{ft_grupo_rama}
\title{Grupos Ramas de actividad económica
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_grupo_rama(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: los datos de la ENFT.
Puede ser conexión a base de datos.}
}
\value{
Los datos suministrados en el argumento tbl con la variables \code{grupo_rama} agregada
}
\description{
Grupos Ramas de actividad económica
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
enft <- ft_grupo_rama(enft)
}
}
