% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_horas_semanal}
\alias{ft_horas_semanal}
\title{Horas trabajadas a la semana
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_horas_semanal(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{horas_semanal} agregada.
}
\description{
En el periodo 2000-2005 se imputó cero (0) para algunos casos que no aplicaban
esta función toma cuenta de esa situación eliminando todos los valores
asignados en cero (0).
}
\examples{
\dontrun{
enft <- ft_horas_semanal(enft)
}
}
