% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_ayuda_gobierno}
\alias{ft_ing_ayuda_gobierno}
\title{Ingreso monetario no laboral por ayuda del gobierno para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_ayuda_gobierno(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} el data.frame con los datos de la ENFT.
Use datos locales; materialice previamente las consultas remotas.}
}
\value{
\link{data.frame} los datos del argumento \code{tbl} con la columna \code{ing_ayuda_gobierno} agregrada.
}
\description{
Ingreso monetario no laboral por ayuda del gobierno para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
enft <- ft_ing_ayuda_gobierno(enft)
}
}
