% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_ext_pension}
\alias{ft_ing_ext_pension}
\title{Ingreso monetario no laboral por pensión del exterior para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_ext_pension(tbl, ing_ext = tbl)
}
\arguments{
\item{tbl}{\link{data.frame} el data.frame con los datos de la ENFT.
Use datos locales; materialice previamente las consultas remotas.}

\item{ing_ext}{\link{data.frame} el data.frame con los datos de la tabla de ingresos externos.
Vea detalles en la función \link{ft_pobreza_monetaria}.}
}
\value{
\link{data.frame} los datos del argumento \code{tbl} con la columna \code{ing_ext_pension} agregrada.
}
\description{
Ingreso monetario no laboral por pensión del exterior para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
enft <- ft_ing_ext_pension(enft, ing_ext)
}
}
