% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_imputado_vivienda_propia}
\alias{ft_ing_imputado_vivienda_propia}
\title{Ingreso no monetario no laboral imputado por uso de vivienda propia para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_imputado_vivienda_propia(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} el data.frame con los datos de la ENFT.}
}
\value{
\link{data.frame} los datos del argumento \code{tbl} con la columna \code{ing_imputado_vivienda_propia} agregrada.
}
\description{
Ingreso no monetario no laboral imputado por uso de vivienda propia para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
enft <- ft_ing_imputado_vivienda_propia(enft)
}
}
