% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_remesas_ext_20001}
\alias{ft_ing_remesas_ext_20001}
\title{Ingreso monetario no laboral por remesas del exterior 2000/1 para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_remesas_ext_20001(ing_ext)
}
\arguments{
\item{ing_ext}{\link{data.frame} el data.frame con los datos de la tabla de ingresos externos.
Vea detalles en la función \link{ft_pobreza_monetaria}.}
}
\value{
\link{data.frame} los datos del argumento \code{ing_ext} con la columna \code{remesas_a} agregrada.
}
\description{
Ingreso monetario no laboral por remesas del exterior 2000/1 para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
