% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_remesas_ext_20002_20162}
\alias{ft_ing_remesas_ext_20002_20162}
\title{Ingreso monetario no laboral por remesas del exterior 2000/2 a 2016/2 para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_remesas_ext_20002_20162(remesas)
}
\arguments{
\item{remesas}{\link{data.frame} el data.frame con los datos de la tabla de remesas.
Vea detalles en la función \link{ft_pobreza_monetaria}.}
}
\value{
\link{data.frame} los datos del argumento \code{remesas} con la columna \code{remesas_b} agregrada.
}
\description{
Ingreso monetario no laboral por remesas del exterior 2000/2 a 2016/2 para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
