% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_pea_abierta}
\alias{ft_pea_abierta}
\title{Población Económicamente Activa (PEA) abierta}
\usage{
ft_pea_abierta(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{pea_abierta} agregada.
}
\description{
Población Económicamente Activa (PEA) abierta
}
\examples{
\dontrun{
enft <- ft_pea_abierta(enft)
}
}
