% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demograficos.R
\name{ft_regiones_desarrollo}
\alias{ft_regiones_desarrollo}
\title{Regiones de desarrollo de la República Dominicana
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_regiones_desarrollo(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} con datos de la base de datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{regiones_desarrollo} adicionada.
}
\description{
Regiones de desarrollo de la República Dominicana
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
  enft <- ft_regiones_desarrollo(enft)
}
}
