% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demograficos.R
\name{ft_regiones_desarrollo_685_00}
\alias{ft_regiones_desarrollo_685_00}
\title{Regiones de desarrollo según decreto 685-00
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_regiones_desarrollo_685_00(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} con los datos de la encuesta}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{regiones_desarrollo_685_00} adicionada.
}
\description{
Regiones de desarrollo según decreto 685-00
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
enft <- ft_regiones_desarrollo_685_00(enft)
}
}
