% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_Labels.R
\name{ft_set_Dict}
\alias{ft_set_Dict}
\alias{ft_set_labels}
\alias{ft_setLabels}
\title{Asigna etiquetas de datos a las variables especificadas}
\usage{
ft_set_Dict(
  tbl,
  dict = NULL,
  subset = NULL,
  ...,
  version = NULL,
  at = NULL,
  con = NULL
)

ft_set_labels(tbl, dict = NULL, vars = NULL)

ft_setLabels(tbl, dict = NULL, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{Dict}: Diccionario con las etiquetas de las variables}

\item{subset}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{...}{argumentos adicionales de \code{labeler::set_Dict}}

\item{version, at, con}{Seleccion opcional de revision; ver \code{\link[=ft_dict]{ft_dict()}}.}

\item{vars}{\link{character}: Alias deprecated de \code{subset} mantenido por compatibilidad.}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Etiquetas de datos \code{vignette("labeler", package = "labeler")}
}
