% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_Labels.R
\name{ft_use_labels}
\alias{ft_use_labels}
\alias{ft_useLabels}
\title{Utiliza las etiquetas de datos de una variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_use_labels(tbl, dict = enftr::dict, vars = NULL)

ft_useLabels(tbl, dict = enftr::dict, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{list}: Diccionario de la base de datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos suministrados en el argumento \code{tbl}, pero en lugar de
valores utilizando las etiquetas de datos correspondientes
}
\description{
Utiliza las etiquetas de datos de una variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  enft <- data.frame(S2_P8 = c(1, 2))
  enft
  ft_use_labels(enft)
}
}
\seealso{
Etiquetas de datos \code{vignette("labels", package = "endomer")}
}
