% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_Labels.R
\name{ft_with_Dict}
\alias{ft_with_Dict}
\alias{ft_use_labels}
\alias{ft_useLabels}
\title{Utiliza las etiquetas de datos de una variable/dataset}
\usage{
ft_with_Dict(tbl, dict = NULL, subset = NULL, ...)

ft_use_labels(tbl, dict = NULL, vars = NULL, ...)

ft_useLabels(tbl, dict = NULL, vars = NULL, ...)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{Dict}: Diccionario con las etiquetas de las variables}

\item{subset}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{...}{argumentos adicionales de \code{labeler::with_Dict}}

\item{vars}{\link{character}: Alias deprecated de \code{subset} mantenido por compatibilidad.}
}
\value{
El dataset modificado
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
