% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demograficos.R
\name{ft_zona}
\alias{ft_zona}
\alias{ft_compute_zona}
\title{Zona de residencia
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_zona(tbl)

ft_compute_zona(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable \code{zona}
adicionada.
}
\description{
La base de datos de ne ENFT en su primera versión imputaba las zonas de
residencia con los valores de 0 y 1, para compatibilidad con la segunda
versión y la encuesta continua (ENCFT) se crea una variable zona que imputa
los valores como 1 y 2.
}
\examples{
(
  enft1 <- data.frame(
    PERIALFA = rep("1/2016", 5),
    S1_P4 = sample(c(0, 1), 5, replace = TRUE)
  )
)
ft_zona(enft1)

(
  enft2 <- data.frame(
    EFT_PERIODO = rep("1/2016", 5),
    EFT_ZONA = sample(c(0, 1), 5, replace = TRUE)
  )
)
ft_zona(enft1)
}
