% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demograficos.R
\name{ft_zona_desarrollo_fronterizo}
\alias{ft_zona_desarrollo_fronterizo}
\title{Zona Especial de Desarrollo Fronterizo de la República Dominicana
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_zona_desarrollo_fronterizo(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} con datos de la base de datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable
\code{zona_desarrollo_fronterizo} adicionada.
}
\description{
Vea Ley 28-01 de la República Dominicana.
}
\examples{
\dontrun{
 enft <- ft_zona_desarrollo_fronterizo(enft)
}
}
