% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poblacion_onaplan}
\alias{poblacion_onaplan}
\title{Población estimada ONAPLAN
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
An object of class \code{data.frame} with 18 rows and 2 columns.
}
\usage{
poblacion_onaplan
}
\description{
Población estimada ONAPLAN
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{datasets}
