% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tdc_oficial}
\alias{tdc_oficial}
\title{Tipos de cambio para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2440 rows and 5 columns.
}
\usage{
tdc_oficial
}
\description{
Tipos de cambio para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{datasets}
