% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_browse_dict.R
\name{ft_browse_dict}
\alias{ft_browse_dict}
\title{Navegar el diccionario de la base de datos
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_browse_dict(dict)
}
\arguments{
\item{dict}{\link{list} diccionario de la base de datos}
}
\value{
widget html interactivo con el diccionario de la encuesta
}
\description{
Navegar el diccionario de la base de datos
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
ft_browse_dict()
}
}
