% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_db_connect.R
\name{ft_db_connect}
\alias{ft_db_connect}
\alias{ft_dbConnect}
\title{Conexión a base de datos}
\usage{
ft_db_connect()

ft_dbConnect()
}
\value{
Conexión a base de datos.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Vea \code{Dmisc::\link[Dmisc:db_connect]{db_connect}}
}
\examples{
\dontrun{
  conn <- ft_db_connect()
}
}
