% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_horas_semanal}
\alias{ft_horas_semanal}
\title{Horas trabajadas a la semana}
\usage{
ft_horas_semanal(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{horas_semanal} agregada.
}
\description{
Horas trabajadas a la semana
}
\examples{
\dontrun{
enft <- ft_horas_semanal(enft)
}
}
