% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_regalos_ext}
\alias{ft_ing_regalos_ext}
\title{Ingreso no monetario no laboral por regalos del exterior para el c치lculo de pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_regalos_ext(tbl, ing_ext)
}
\arguments{
\item{tbl}{\link{data.frame} el data.frame con los datos de la ENFT.
Puede ser conexi칩n a base de datos.}

\item{ing_ext}{\link{data.frame} el data.frame con los datos de la tabla de ingresos externos.
Vea detalles en la funci칩n \link{ft_pobreza_monetaria}.}
}
\value{
\link{data.frame} los datos del argumento \code{tbl} con la columna \code{ing_regalos_ext} agregrada.
}
\description{
Ingreso no monetario no laboral por regalos del exterior para el c치lculo de pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
enft <- ft_ing_regalos_ext(enft, ing_ext)
}
}
