% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_total_pobreza_monetaria}
\alias{ft_ing_total_pobreza_monetaria}
\title{Ingreso total hogares para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_ing_total_pobreza_monetaria(
  tbl,
  ing_ext = tbl,
  remesas = tbl,
  .keep = FALSE,
  .reuse = FALSE
)
}
\arguments{
\item{tbl}{\link{data.frame}: datos de la ENFT. Puede ser conexión a base de datos.}

\item{ing_ext}{\link{data.frame}: datos de ingresos del exterior. Vea detalles.}

\item{remesas}{\link{data.frame}: datos de remesas. Vea detalles.}

\item{.keep}{indica si se deben mantener las variables intermedias en la data.
Puede ser \code{TRUE} o \code{FALSE} (default), o un vector de characeter con las variables a mantener.}

\item{.reuse}{indica si se deben reutilizar las variables intermedias que estén disponibles en la data.
Puede ser \code{TRUE} o \code{FALSE}(default). O un vector de characeter con las variables a reutilizar.}
}
\value{
\link{data.frame}: los datos del argumento \code{tbl} con la variable \code{ing_total_pobreza_monetaria} agregada.
}
\description{
Ingreso total hogares para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
enft <- ft_pobreza_monetaria(enft, enft_ing_ext, enft_remesas)
}
}
