% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_ing_utilidades_empresariales}
\alias{ft_ing_utilidades_empresariales}
\title{Ingreso monetario laboral por utilidades empresariales para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_ing_utilidades_empresariales(tbl)
}
\arguments{
\item{tbl}{\link{data.frame} el data.frame con los datos de la ENFT.
Puede ser conexión a base de datos.}
}
\value{
\link{data.frame} los datos del argumento \code{tbl} con la columna \code{ing_utilidades_empresariales} agregrada.
}
\description{
Ingreso monetario laboral por utilidades empresariales para el cálculo de la pobreza monetaria
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
enft <- ft_ing_utilidades_empresariales(enft)
}
}
