% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_ingreso_laboral_mensual}
\alias{ft_ingreso_laboral_mensual}
\title{Ingreso laboral mensual
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
ft_ingreso_laboral_mensual(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{ingreso_laboral_mensual} agregada.
}
\description{
Ingreso laboral mensual
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
enft <- ft_ingreso_laboral_mensual(enft)
}
}
