% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_ocupado}
\alias{ft_ocupado}
\title{Población ocupada}
\usage{
ft_ocupado(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{ocupado} agregada.
}
\description{
Población ocupada
}
\examples{
\dontrun{
 enft <- ft_ocupado(enft)
}
}
