% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercado-laboral.R
\name{ft_poblacion_inactiva}
\alias{ft_poblacion_inactiva}
\title{Población inactiva (No PEA)}
\usage{
ft_poblacion_inactiva(tbl, min_edad = 15)
}
\arguments{
\item{tbl}{\link{data.frame} Data.frame con los datos de la encuesta}

\item{min_edad}{\link{integer} Edad mínima para considerar una persona en edad de trabajar}
}
\value{
\link{data.frame} los datos suministrados en el input tbl, pero con la
variable \code{poblacion_inactiva} agregada.
}
\description{
Población inactiva (No PEA)
}
\examples{
\dontrun{
enft <- ft_poblacion_inactiva(enft)
}
}
