% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociales.R
\name{ft_pobreza_monetaria}
\alias{ft_pobreza_monetaria}
\title{Cálculo de pobreza monetaria con metodología oficial
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ft_pobreza_monetaria(conn, name = "pobreza_monetaria")
}
\arguments{
\item{conn}{conexión a la base de datos}

\item{name}{nombre de la tabla en la base de datos}
}
\value{
data.frame con los datos de la tabla
}
\description{
Cálculo de pobreza monetaria con metodología oficial
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Por el momento, carga los datos desde una tabla
en la base de datos.
}
\examples{
\dontrun{
  pobreza_monetaria <- pobreza_monetaria(conn, "pobreza_monetaria")
}
}
