% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_compute_zona.R
\name{ft_zona}
\alias{ft_zona}
\alias{ft_compute_zona}
\title{Zona de residencia}
\usage{
ft_zona(tbl)

ft_compute_zona(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}
}
\value{
Los datos suministrados en el input \code{tbl} con la variable \code{ano}
adicionada.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
La base de datos de ne ENFT en su primera versión imputaba las zonas de
residencia con los valores de 0 y 1, para compatibilidad con la segunda
versión y la encuesta continua (ENCFT) se crea una variable zona que imputa
los valores como 1 y 2.
}
\examples{
  (enft <- data.frame(EFT_PERIODO = rep("1/2016", 5), S1_P4 = sample(c(0, 1), 5, replace = TRUE)))
  ft_zona(enft)
}
